[IMP] work in progress start to split the voodoo in many file with customisable hook for the service used like odoo, wagon, rails, ...
