+++ voodoo/__init__.py
@@ -1,5 +1,8 @@
 #!/usr/bin/env python
 # coding: utf-8
 from .main import Voodoo
+from . import hook
+from .odoo import hook
+from .wagon import hook
 
 __version__ = '2.0.2'
